Remove ConfigureLogging and UseServer(IServer) from IWebApplicationBuilder
